Further simplify signature verification code.
It turns out that at least recent Rust compilers are able to correctly
unify the types of match arms to a particular trait type if the result
of the match is assigned to a variable that has been given an explicit
type annotation. This means we can make the code even neater.