Convert Repository types to use interior mutability
This changes `Repository` to use interior mutability
in order to store state. At the moment, `EphemeralRepository`
is the only repository type that stores state. By switching
over to internally using `Arc<Mutex<...>>>` to wrap the state,
it allows repositories to be used concurrently across threads.
This trait will be especially important as we transition
over to using asynchronous io with Hyper 0.11+.