Remove duplicate code when creating client
This block of code is either redundant, or was supposed to
fetch the metadata from the remote server if it wasn't present
locally. The documentation suggests that it will only try
loading from the local server, so I erred on just removing
this chunk of code, rather than doing the less-safe thing
of downloading the initial metadata from an unverified
server.