Atomically write metadata to file system
Before this patch, `FileSystemRepository::store_metadata` would
stream writes into the destination. If the system crashed,
it's possible the metadata could be truncated. This refactors
the code to atomically write metadata to avoid this happening.

Note that this changes the behavior of
`FileSystemRepository::store_target`. It used to store the
temporary files in a `$REPO/temp` directory, but instead
this moves the temp files into the same directory as the
destination file. This was done to avoid the case where
`temp/` was stored on a separate filesystem than the
destination. If that were the case,
`NamedTemporaryFile::persist` would not be atomic.