Clients should locally save metadata
This patch merges `Client::update_local` and
`Client::update_remote` in order to update `Tuf` following the
pattern as described in the
[spec section 5.1](https://github.com/theupdateframework/specification/blob/master/tuf-spec.md#the-client-application).

One potential problem though is I'm not sure how we can atomically
update the local repository and the `Tuf` object, so for ease of
implementation, this patch will update `Tuf` and could potentially
error out writing the metadata to the local repository. We could
instead implement some sort of rollback mechanism to `Tuf`, but
I figure we can do that later if it's necessary.

Closes #153