Port over to futures and async/await
This is an experiment on porting rust-tuf over to using futures
and async/await, which is what we use on Fuchsia. This is in contrast
with #151, which uses futures 0.1, which is compatible with stable
rust, but dramatically more invasive of a change, which is why
I haven't rebased it on top of remotes/origin/develop yet.

The downside of this is that landing this patch would make rust-tuf
require nightly until async/await is stablized. Unfortunately there
is no clear date on when that is going to occur.