Upgrade to futures-preview 0.3.0-alpha.15
This updates rust-tuf to futures-preview 0.3.0-alpha.15, which
required adding a number of new constraints in a bunch of
places due to a change in AsyncRead. In addition, futures-preview
grew BoxFuture, which replaces the need for us to use TufFuture.
Finally, this changes EphemeralRepository to not have to copy
the target files when calling `.fetch_target`.

Change-Id: Ia20f50ad4c563f009a8375f4cde03914827911e5