Import fuchsia go-tuf metadata
We want to make sure that rust-tuf is compatible with Fuchsia's go-tuf
fork, and eventually other TUF implementations. This patch starts that
process by importing the Fuchsia [go-tuf] test metadata into the
rust-tuf repository in order to make sure rust-tuf can parse the
metadata. There are two sets of metadata:

* go-tuf-5527fe: what the library currently generates.
* go-tuf-transition-M4: what the library will eventually generate once
  all the patches have landed.

In order to ease review, this patch only includes the metadata, and a
separate patch will add the integration test.

[go-tuf]: https://fuchsia.googlesource.com/third_party/go-tuf/+/refs/heads/master/client/testdata/