delegation keys are a map of keyid to key; allow unknown key ids
Like in #184, the [spec] in section 4.5, the `"keys"` object is a map
of key id to key. In addition, this changes Delegations::new to take
the keys as a HashMap instead of computing the keyid from the key.
This is to allow rust-tuf to ignore keys computed with alternative
algorithms. For example, python-tuf supports the ability to optionally
generate key ids with `hexdigest(sha512(cjson(public_key)))`. rust-tuf
should allow (and right now, ignore) those keys.

[spec]: https://github.com/theupdateframework/specification/blob/master/tuf-spec.md

Change-Id: Ifa9a2ab3e4342d70d730e432e38eee6cbca5f1dc