Add support for custom TargetDescription metadata
This extends TargetDescription to capture the `custom` field as a map
of `String`s to `serde_json::Value`s. This ties the interchange more
closely with the cjson format, but this could be addressed once more
interchange formats are implemented.

In addition, this adds a new function `Client::fetch_target_description`,
to allow end users to discover this target information.