re-enable building on stable rust, switch to futures 0.3.1
With the Rust 1.39.0 release, async-await is now stablized so we can start
testing on stable rust again. This also switches over to futures-0.3.1,
which should be semver compatible with all 0.3.x changes.

Finally, it switches away from depending on the `futures` crate, and
instead depends on `futures-io` and `futures-util` (and `futures-executor`
in tests) to reduce crate download size.

Change-Id: I5c939de734bf3f8903895d43d9e85a14e2c71ff6