Add a json pretty printing interchange
This adds an interchange called `JsonPretty`, which makes it easier for
humans to review metadata changes. It is identical to
`tuf::interchange::Json`, except for the implementation of `to_writer`.
Instead of writing out the canonical form, it writes out a pretty printed
form of the data.

In addition, it changes `SnapshotMetadataBuilder::insert_metadata_with_path`
and `TimestampMetadataBuilder::from_snapshot` to use `D::to_writer`. This
makes sure that the hash of these metadata files are computed from the
format written to disk, instead of the canonical format.