Enable IPv6 zone ID support for HttpRepository
Accomplished by switching to `http::Uri` rather than `url::Url`. This
is somewhat unfortunate, as the ergonomics of `http::Uri` are notably
worse, but the result is that zone IDs work.

For API compatibility purposes, the changes manifest as a new struct
called HttpRepository2 and HttpRepositoryBuilder2. In a subsequent
commit we can delete HttpRepository and rename HttpRepository2.