Disable cargo-audit on pull requests
This check is failing on https://github.com/heartsucker/rust-tuf/pull/259 , and I don't think that's the right place to be surfacing this warning. It also looks like it automatically cancels all the other workflows when this happens, which is unhelpful.

It's mildly useful info to know that spin's maintainer is gone, but completely un-actionable while writing or reviewing pull requests on rust-tuf.