Fixup io::Error not_found conversion
When a Client is configured with a local FileSystemRepository, it is
expected that it does not contain metadata until it is fetched from a
remote repository. The client initializer will first check the local
repository for root.json metadata before checking with the remote
repository, but it will only try the remote repository if the local
fetch_metadata fails with Error::NotFound and the current implementation
of FileSystemRepository produces an Error::Opaque when the requested
metadata does not exist.

This change maps io errors with kind() == io::ErrorKind::NotFound to
Error::NotFound. and adds a test for a blank FileSystemRepository to
show it produces the expected error.