Split read/write half of Repository trait
A TUF client has no need to store metadata in a remote repository, and
some repository implementations cannot implement the store_* methods, so
this change splits defines the Repository trait to be a supertrait of
RepositoryProvider (the read half) and RepositoryStorage (the write
half).  Other than requiring clients to import additional traits, this
change should not change any existing functionality.