Metadata trait methods preserve original bytes
The current definition the fetch_metadata and store_metadata repository
trait methods require the repository trait implementations to parse the
metadata, and the SignedMetadata structs and DataInterchange traits
cannot provide a guarantee that parsing and re-serializing metadata will
produce byte-for-byte identical metadata.

This behavior is an issue for TUF as during the metadata update
workflow, clients fetch metadata from a remote repository and store it
in the local repository for later use. However, for metadata referenced
by snapshot, the hash of the data must match in order to use the local
version.

This change:
 * modifies the fetch_metadata and store_metadata trait methods to
   interact with AsyncRead instead of SignedMetadata so that the
   original unparsed bytes can be stored as-is in the local repository
   as long as the metadata is determined to be valid.
 * defines the RawSignedMetadata type, which is a simple wrapper around
   raw bytes and contains type information identifying its inner
   Metadata type and serialization format.
 * introduces a new type to wrap an instance of the Repository traits
   and provide an API surface that can store the raw metadata and fetch
   both raw and parsed metadata to simplify interacting with the now
   less ergonomic trait methods.
 * moves the metadata max length checks and hash checks out of the
   repository trait implementations and into the RepositoryClient,
   ensuring invalid metadata provided by external RepositoryProviders
   will not be trusted (for metadata that is is hash-checked).