Allow Client builders to accept Repository values
This change modifies the local/remote parameters for Client's
constructors to be `impl Into<Repository<R, D>>`, which both preserves
the existing behavior where implementations of RepositoryProvider and/or
RepositoryStorage are accepted and allows callers to provide a
Repository directly.

As long as one of the parameters is a Repository instance, there won't
be a need to turbofish or annotate the Client type to include the
DataInterchange parameter.