Add a test helper to create repository metadata
This adds a proof-of-concept RepoBuilder builder, which is a builder to
simplify the process of defining repository metadata. Some of tests in
`src/client.rs` were migrated to this builder, which greatly reduced the
size of the functions.

The API is still a little rough though, so it is currently marked as a
private `#[cfg(test)]` crate. Before it's exposed, we need to at least:

* Fetch and validate the old metadata before basing new metadata off of
  it.
* Don't overwrite metadata that already exists.
* Auto-increment the version numbers.
* Optionally allow the timestamp version number to be generated off of
  the current UTC time, to ease working with test repositories.
* Merge targets and delegates.
* Come up with a better way to handle delegation metadata.
* Add documentation and examples.
* Make a proper error for when `repo_builder::sign` is given an empty
  vec of private keys.

Change-Id: I78e20aa2d44448a5c96520aa7bbd808fa354ed87