Control when metadata is deserialized
This rewrites rust-tuf to be more strict on when metadata is deserialized.
This is to start laying the groundwork to prevent rust-tuf from being
vulnerable to parser exploits (like [CVE-2017-18349]). Almost all
deserialization is moved into the new `verify.rs` module, which validates
the metadata's signature before returning the deserialized metadata to the
caller.

There are two remaining cases where metadata is parsed outside of `verify`:

* `Client::from_trusted_local`, where we don't have any keys to establish
  trust of the initial root metadata.
* `Client::update_root`, where we fetch the latest root metadata, and walk
  from the current trusted metadata version to the latest root metadata
  version.

These are explicitly called out with WARNING comments.

Also, this starts incorporating the TUF workflow spec into the `Tuf`
functionality, in order to help reviewers understand which code is
implementing which part of the spec, or make it obvious when we diverge
from it.

Bug: #291

[CVE-2017-18349]: https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2017-18349