Remove tuf::Client::with_trusted_root_keyids
In https://github.com/theupdateframework/taps/blob/master/tap12.md, the
TUF community is considering removing the requirement that the keyid must
be derived from the public key. If this lands, then it wouldn't be safe to
initialize trust with just the keyids. To prepare for this transition,
this patch removes `tuf::Client::with_trusted_root_keyids`. If this
functionality is still needed, it'd be relatively straightforward for a
client to re-write this method in their codebase.

In help implement this, this patch also extends RootMetadata to add helper
methods to extract out role public keys.