Persist unversioned metadata to local store
The TUF-1.0.9 spec states that the metadata should be written to
non-volatile storage as an unversioned path after verification.

This also keeps the extension where we also persist the versioned root
metadata. This allows a client to be initialized first from an trusted root
metadata (which could be stored in read-only storage), then load metadata
from the local store, and finally then the remote metadata.

Change-Id: Id62685db069b65e094666c97464583df4c2a1163