Make RepositoryStorage object-safe
This changes the `RepositoryStorage` type to be object-safe by passing in
the `AsyncRead` argument as a `&mut dyn AsyncRead`. Furthermore, Rust
doesn't allow for a type like `Box<dyn A + B>`. So in order to make the
`Repository` struct take a dyn boxed repository, this patch also adds a
trait `RepositoryStorageProvider`, which has as a super trait
`RepositoryStorage` and `RepositoryProvider`.

Test: Updated all the tests to work with this, and added a test to make
sure these traits stay object-safe.

Change-Id: I33c014ae07ab4cf409aa6832e208098354f5423b