Cite TUF spec, move timestamp expiry check to end.
This adds direct citations of the TUF-1.0.5 spec in the `Tuf` client to
make it easier for reviewers to directly see how each code block implements
the spec, or how we diverge from it.

It makes one minor code change. Previously the timestamp expiration was
performed after signature verification. This moves the expiration check
to be the last check to match the spec.

Change-Id: I4125ff4d2e40abdbb177c03665d591b5c90e1794