Client::update should fail if we can't write metadata
Previously, `Client::update` would warn, but ultimately ignore, if we
failed to persist metadata to the local store. This doesn't play well with
trying to use TUF offline.

Consider the case where a client is doing an update for some new targets
file. If we do an update, but we run out of space before writing the new
targets file. If we reboot the device and lose our connection to the
repository, we can no longer validate the old timestamp file.

This patch changes the client to error when this happens. This would allow
a user of `tuf::Client` to handle this error (like clearing some files from
local storage before re-trying the update). If a user instead wants the old
behavior, they could initialize a client with a local repository that only
warns if an error occurs.

Test: This adds a test case where where the client will err out if we fail
to write an update.

Change-Id: I31ffa2fc98353a16e73adeec6eb8da729784db92