Initialize clients with local metadata
This updates the `tuf::Client` constructors to load the initial metadata
from the local repository. It does this by refactoring `Client` to support
updating from arbitrary repositories, and optionally writing the metadata
back to the local repository (since during initialization we don't want to
write back what we just read). One thing to note is that all but the
root metadata is stored unversioned, so during initialization we are
treating things as if consistent_snapshots is false.

In order to do this, this also changes
`tuf::Repository::store_{metadata,targets` to take `&self`.

Test: This adds a few tests to check this behavior.
* Check that using all client constructors will load metadata during
  initialization.
* Check that we ignore errors if metadata is missing  expired, or
  malformed.
* Check that we won't try to load metadata from the local repository after
  client initialization.

Change-Id: I1a1e4aa1ef8a644ddd47671f34bbf091d0b92eb9