Make `Client::update_root()` public
Users of `Client::trusted_root()` might only care about the latest root
metadata, and none of the other metadata, but unfortunately the only way
to get get the latest root is to call `Client::update()`, which downloads
all the other metadata. This can be a problem because it's likely that the
root metadata has a long expiration, but the other metadata may have expired.

To address this, this exposes `Client::update_root()`, which allows these
users to just make sure the root metadata is up to date, and avoid the risk
of `Client::update()` failing with expired metadata.

Change-Id: Ia1ce171828b322f90899968a5e3f593143e89126