Put RSA behind a feature flag
RSA key support doesn't conform to the spec yet, so put it behind an
`unstable_rsa` feature flag. This will allow us to eventually conform to
the spec without making a breaking API change.

In addition, this marks all the key enums as #[non_exhaustive], so we
can add new keys in the future without breaking our public APIs.