Remove tuf::client::PathTranslator
This removes tuf::client::PathTranslator. I'm not sure if this is a
holdover from pre TUF-1.0 era, but as of right now I'm not aware of any
strong use cases for translating paths. The only one I can come up with is
if some repository wants to escape OS-specific characters, but that could
be done in a repository wrapper, or in the FileSystemRepository itself.