Optional metafile length and snapshot
According to the [spec] it is optional for timestamp to contain the length
and hash of the snapshot, and for snapshot to contain the length and hash
for targets. This is considered safe because the metadata still needs to
have it's metadata signature verified before used. However, the client
must set an upper bound on the download size if the metadata does not
specify the download length.

The upper bounds have been updated to sync with python-tuf's [defaults].

[spec]: https://theupdateframework.github.io/specification/latest/#metapath-hashes
[defaults]: https://github.com/theupdateframework/python-tuf/blob/30ba6e9f9ab25e0370e29ce574dada2d8809afa0/tuf/settings.py