Validate all hashes; try all target hash prefixes
This changes how the client downloads metadata and targets. Previously we
would only make one attempt to download the most preferred hash algorithm.
However, according to [update-snapshot] and [update-targets], we should
verify all hashes are valid. This is done by moving much of the hash logic
into `tuf::repository::Repository`.

Furthermore, [fetch-target] says that when we are downloading a target
with consistent snapshots are enabled, we will now attempt to download the
target for each hash prefix.

[update-snapshot]: https://theupdateframework.github.io/specification/v1.0.26/#update-snapshot
[update-targets]: https://theupdateframework.github.io/specification/v1.0.26/#update-targets
[fetch-target]: https://theupdateframework.github.io/specification/v1.0.26/#fetch-target