Add ephemeral and file system repo transactions
This introduces the notion of a repository transaction, and implements it
for EphemeralRepository and FileSystemRepository. These are special
repositories that are designed to be used when creating new TUF metadata.
Rather than writing directly to the repository, these stage any queued up
metadata in a staging location, and allows users to as-atomic-as-possible
write that data back to the source repository when everything has been
staged.

One caveat. While the FileSystemTransaction can atomically write each
file, it may be possible for a system error to cause `commit()` to fail
partway through writing files.