Optimize MetadataPath for core metadata types
This changes MetadataPath to wrap a `Cow<'static, str>`, rather than a
`String`, which allows us to avoid an allocation for our most common types.
It also adds helpers for making `MetadataPath` from the core metadata
types.

In addition, this changes the error types `ExpiredMetadata` and
`MissingMetadata` to wrap a `MetadataPath` rather than a `Role`. We'll
eventually use this to return proper errors during delegated target
resolution.