Implement update start time
This patch implements [TUF 1.0.30 section 5.1.1], where we record the
update start time at the beginning of the client update workflow, and
compare all expirations to this. This allows us to avoid the risk of the
metadata expiring during the update process.

In addition, this adds `Client::update_with_start_time()`, which allows
the client to specify the start time to use, rather than the current
time. This allows the client to work with expired metadata (and so there
is a warning in the documentation), it could also be used to share
expiration time with TUF and other external processes.

[TUF 1.0.30 section 5.1.1]: https://theupdateframework.github.io/specification/latest/#fix-time