Builder should copy targets and delegations by default
When creating a new targets metadata, we should make sure that we copy over
the old targets and delegations by default. This can be overridden with the
new `inherit_from_trusted_targets(bool)` method.

This also renames the old `inherit_targets` to
`inherit_from_trusted_snapshot` to be more clear.