Switch RepositoryStorage::store_* to take &self
During our migration to futures, we switched
RepositoryStorage::store_metadata and RepositoryStorage::store_targets to
take `&mut self` to make it easier to reason about changes. This change
however prevented us from concurrently changing the repository, which can
speed up large commits. This change switches the repositories back to
use `&self` to enable this.

One consequence of doing this is we need to change the batch update
implementation `commit()` to now be fallible. This is implemented by
incrementing a version number on every repository change, and storing the
initial version number in the batch. We have a conflict if the versions
don't match during commit.