Add support for time versioning
Time versioning is a metadata versioning approach where we use the
current UTC unix timestamp when creating non-root metadata, rather than
monotonically increasing the metadata version.

This is useful when working with a test device that may have TUF
metadata created from one source, then updated with developer-created
metadata. Assuming the user's workstation and test device have their
system times in sync, the user can create new TUF metadata that is a
newer version than the metadata on the device without needing to query
the device to see what it thinks is the latest metadata version.