Rename pouf::Json to Pouf1
After a discussion with the TUF community, we decided it made more sense
to name our implementation of [POUF-1] as `Pouf1`, not `Json`, because it
has additional restrictions on top of JSON. Furthermore, it doesn't make
sense to call this [CJSON] since TUF doesn't allow for raw bytes in
strings.

This also moves `JsonPretty` into the interop-tests, since this is only
used to make the interop tests easier to review. I've filed [161] with
the TUF spec to consider allowing whitespace in POUF-1. If this is accepted
we can move this back into the tuf::pouf submodule.

[POUF-1]: https://github.com/theupdateframework/taps/blob/master/POUFs/reference-POUF/pouf1.md
[161]: https://github.com/theupdateframework/taps/issues/161
[CJSON]: https://wiki.laptop.org/go/Canonical_JSON