Increase minimum Rust version tested in CI to 1.63.0.
The latest version of tempfile (because of its dependency rustix, at
least) requires 1.63.0. Similarly, the latest version of Tokio
requires it. The latest Chrono also requires 1.57.0.

Add a clippy.toml that prevents Clippy from recommending changes that
require a newer toolchain. This is required or else the `clippy` job
fails with advice to use APIs that are only in later versions.